Edit temp read method

+++ src/main/java/IMUmanager.java
@@ -58,7 +58,7 @@ public class IMUmanager {
 
     private static void writeMagReg( int register, byte value ){
         try {
-            System.out.println("writing to mag register: "+ register+ " value: "+value);
+            System.out.println("writing to mag register: "+ register+ " value: "+(int)value);
            bgi_mag.write(register, value);
         } catch (IOException e) {
             System.out.println("Failed to write to Mag Register");
